move app creation into own function
